Refactor some admin auth-related functions into a separate import
Reusing a single implementation of these functions seems better than
duplicating them, and explicitly importing them seems better than
relying on a magic global.
